eeg_fm/eeg_foundation_models/requirements.txt
torch
lightning
mne
numpy
einops
pyyaml
scikit-learn
torchmetrics
moabb
vector-quantize-pytorch
transformers
sentencepiece
